Rename PotionEffect.java to ConsumableEffect.java
