+++ Server/src/main/java/Server/plugin/consumable/effects/MultiEffect.java
@@ -1,16 +1,16 @@
 package plugin.consumable.effects;
 
 import core.game.node.entity.player.Player;
-import plugin.consumable.PotionEffect;
+import plugin.consumable.ConsumableEffect;
 
-public class MultiEffect extends PotionEffect {
-    private PotionEffect[] effects;
-    public MultiEffect(PotionEffect... effects){
+public class MultiEffect extends ConsumableEffect {
+    private ConsumableEffect[] effects;
+    public MultiEffect(ConsumableEffect... effects){
         this.effects = effects;
     }
     @Override
     public void activate(Player p) {
-        for(PotionEffect e : effects){
+        for(ConsumableEffect e : effects){
             e.activate(p);
         }
     }
