+++ Server/src/main/java/Server/plugin/consumable/effects/PrayerEffect.java
@@ -1,10 +1,10 @@
 package plugin.consumable.effects;
 
 import core.game.node.entity.player.Player;
-import plugin.consumable.PotionEffect;
+import plugin.consumable.ConsumableEffect;
 import plugin.skill.Skills;
 
-public class PrayerEffect extends PotionEffect {
+public class PrayerEffect extends ConsumableEffect {
     double base, bonus;
     public PrayerEffect(double base, double bonus){
         this.base = base;
